+++ components/shared/constellation/Cargo.toml
@@ -18,9 +18,11 @@ webgpu = ["wgpu-core"]
 [dependencies]
 base = { workspace = true }
 canvas_traits = { workspace = true }
+compositing_traits = { workspace = true }
 devtools_traits = { workspace = true }
 embedder_traits = { workspace = true }
 euclid = { workspace = true }
+fonts_traits = { workspace = true }
 http = { workspace = true }
 hyper_serde = { workspace = true }
 ipc-channel = { workspace = true }
