+++ components/shared/constellation/from_script_message.rs
@@ -14,6 +14,7 @@ use base::id::{
     MessagePortRouterId, PipelineId, ServiceWorkerId, ServiceWorkerRegistrationId, WebViewId,
 };
 use canvas_traits::canvas::{CanvasId, CanvasMsg};
+use compositing_traits::CrossProcessCompositorApi;
 use devtools_traits::{DevtoolScriptControlMsg, ScriptToDevtoolsControlMsg, WorkerId};
 use embedder_traits::{
     AnimationState, EmbedderMsg, FocusSequenceNumber, JSValue, JavaScriptEvaluationError,
@@ -21,13 +22,14 @@ use embedder_traits::{
     WebDriverMessageId,
 };
 use euclid::default::Size2D as UntypedSize2D;
+use fonts_traits::SystemFontServiceProxySender;
 use http::{HeaderMap, Method};
 use ipc_channel::ipc::{IpcReceiver, IpcSender};
 use malloc_size_of_derive::MallocSizeOf;
 use net_traits::policy_container::PolicyContainer;
 use net_traits::request::{Destination, InsecureRequestsPolicy, Referrer, RequestBody};
 use net_traits::storage_thread::StorageType;
-use net_traits::{CoreResourceMsg, ReferrerPolicy, ResourceThreads};
+use net_traits::{ReferrerPolicy, ResourceThreads};
 use profile_traits::mem::MemoryReportResult;
 use profile_traits::{mem, time as profile_time};
 use serde::{Deserialize, Serialize};
@@ -205,8 +207,12 @@ pub struct DOMMessage {
 pub struct SWManagerSenders {
     /// Sender of messages to the constellation.
     pub swmanager_sender: IpcSender<SWManagerMsg>,
-    /// Sender for communicating with resource thread.
-    pub resource_sender: IpcSender<CoreResourceMsg>,
+    /// [`ResourceThreads`] for initating fetches or using i/o.
+    pub resource_threads: ResourceThreads,
+    /// [`CrossProcessCompositorApi`] for communicating with the compositor.
+    pub compositor_api: CrossProcessCompositorApi,
+    /// The [`SystemFontServiceProxy`] used to communicate with the `SystemFontService`.
+    pub system_font_service_sender: SystemFontServiceProxySender,
     /// Sender of messages to the manager.
     pub own_sender: IpcSender<ServiceWorkerMsg>,
     /// Receiver of messages from the constellation.
